Change output data type

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	long long int ll;
 	float f;
 
-	printf("Size of a char: %lu bytes(s)\n",(unsigned long)sizeof(c));
-	printf("Size of an int: %lu bytes(s)\n",(unsigned long)sizeof(i));
-	printf("Size of a long int: %lu bytes(s)\n",(unsigned long)sizeof(l));
-	printf("Size of a long long int: %lu bytes(s)\n",(unsigned long)sizeof(ll));
-	printf("Size of a float: %lu bytes(s)\n",(unsigned long)sizeof(f));
+	printf("Size of a char: %lu bytes(s)\n", (unsigned int)sizeof(c));
+	printf("Size of an int: %lu bytes(s)\n", (unsigned int)sizeof(i));
+	printf("Size of a long int: %lu bytes(s)\n", (unsigned int)sizeof(l));
+	printf("Size of a long long int: %lu bytes(s)\n", (unsigned int)sizeof(ll));
+	printf("Size of a float: %lu bytes(s)\n", (unsigned int)sizeof(f));
 	return (0);
 }
